// src/curseforge/openapi/types.ts
export enum FileRelationType {
    EmbeddedLibrary = 1,
    OptionalDependency = 2,
    RequiredDependency = 3,
    Tool = 4,
    Incompatible = 5,
    Include = 6
}
export enum FileReleaseType {
    Release = 1,
    Beta = 2,
    Alpha = 3
}
// ...
export enum HashAlgo {
    Sha1 = 1,
    Md5 = 2
}